Start of YCbCr, changed to float for the imagejk

+++ image.py
@@ -2,11 +2,11 @@ from skimage import io
 
 def open_image_as_ndarray(filename):
     img = io.imread(filename)
-    return img
+    return img.astype('float')
 
 def get_test_image():
     # RECAP : Opening the image and splitting it into blocks
-    img = io.imread('input_image.png')
+    img = io.imread('input_image.png').astype('float')
     return img
